float to double

+++ TP1/src/geom/AABB.java
@@ -43,7 +43,7 @@ public class AABB extends OBB {
 	 * @throws IllegalArgumentException
 	 *             If the width or the height are negative or null.
 	 */
-	public AABB(float width, float height) throws IllegalArgumentException {
+	public AABB(double width, double height) throws IllegalArgumentException {
 		super(width, height, 0);
 	}
 
@@ -62,8 +62,8 @@ public class AABB extends OBB {
 	 * @throws IllegalArgumentException
 	 *             If the width or the height are negative or null.
 	 */
-	public AABB(Position position, float width, float height) throws IllegalArgumentException {
+	public AABB(Position position, double width, double height) throws IllegalArgumentException {
 		super(position, width, height, 0);
 	}
-	
+
 }
