+++ TP1/src/geom/OBB.java
@@ -9,13 +9,13 @@ package geom;
 public class OBB extends Shape {
 
 	/** The width of the box. */
-	protected float width;
+	protected double width;
 
 	/** The height of the box. */
-	protected float height;
+	protected double height;
 
 	/** The angle of the box. */
-	private float angle;
+	private double angle;
 
 	/**
 	 * OBB constructor with a default width of 10, height of 10 and angle of 0
@@ -57,7 +57,7 @@ public class OBB extends Shape {
 	 * @throws IllegalArgumentException
 	 *             If the width or the height are negative or null.
 	 */
-	public OBB(float width, float height, float angle) throws IllegalArgumentException {
+	public OBB(double width, double height, double angle) throws IllegalArgumentException {
 		super();
 		if (width <= 0)
 			throw new IllegalArgumentException("width is negative or null: " + width);
@@ -88,7 +88,7 @@ public class OBB extends Shape {
 	 * @throws IllegalArgumentException
 	 *             If the width or the height are negative or null.
 	 */
-	public OBB(Position position, float width, float height, float angle) throws IllegalArgumentException {
+	public OBB(Position position, double width, double height, double angle) throws IllegalArgumentException {
 		super(position);
 		if (width <= 0)
 			throw new IllegalArgumentException("width is negative or null: " + width);
@@ -128,7 +128,7 @@ public class OBB extends Shape {
 	 *
 	 * @return The width.
 	 */
-	public float getWidth() {
+	public double getWidth() {
 		return width;
 	}
 
@@ -138,7 +138,7 @@ public class OBB extends Shape {
 	 * @param width
 	 *            The new width.
 	 */
-	public void setWidth(float width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
 
@@ -147,7 +147,7 @@ public class OBB extends Shape {
 	 *
 	 * @return The height.
 	 */
-	public float getHeight() {
+	public double getHeight() {
 		return height;
 	}
 
@@ -157,7 +157,7 @@ public class OBB extends Shape {
 	 * @param height
 	 *            The new height.
 	 */
-	public void setHeight(float height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
 
@@ -166,7 +166,7 @@ public class OBB extends Shape {
 	 *
 	 * @return The angle.
 	 */
-	public float getAngle() {
+	public double getAngle() {
 		return angle;
 	}
 
@@ -176,7 +176,7 @@ public class OBB extends Shape {
 	 * @param angle
 	 *            The new angle.
 	 */
-	public void setAngle(float angle) {
+	public void setAngle(double angle) {
 		this.angle = angle;
 	}
 
@@ -185,8 +185,8 @@ public class OBB extends Shape {
 	 * 
 	 * @return The diagonal length.
 	 */
-	public float getDiagLength() {
-		return (float) Math.sqrt(this.height * this.height + this.width * this.width);
+	public double getDiagLength() {
+		return Math.sqrt(this.height * this.height + this.width * this.width);
 	}
 
 	/**
@@ -194,7 +194,7 @@ public class OBB extends Shape {
 	 * 
 	 * @return The area.
 	 */
-	public float getArea() {
+	public double getArea() {
 		return this.width * this.height;
 	}
