+++ TP1/src/geom/Shape.java
@@ -31,7 +31,7 @@ public abstract class Shape {
 	 * @param y
 	 *            The wanted position for the y axis.
 	 */
-	public Shape(float x, float y) {
+	public Shape(double x, double y) {
 		this.position = new Position(x, y);
 	}
 
@@ -111,7 +111,7 @@ public abstract class Shape {
 	 * @param y
 	 *            The new wanted value for the y axis.
 	 */
-	public void setPosition(float x, float y) {
+	public void setPosition(double x, double y) {
 		this.position = new Position(x, y);
 	}
